Store storage targets, rather than ranks
This means we can now actually support 56 storage targets and that the
locations stored in the database will probably be more useful when
rebuilding.

+++ bp-parity-gen/file_info_hash.c
@@ -23,7 +23,7 @@ void fih_term(FileInfoHash *fih)
     free(fih);
 }
 
-int fih_add_info(FileInfoHash *fih, char *key, uint16_t src, uint64_t size, uint64_t time)
+int fih_add_info(FileInfoHash *fih, char *key, int src, uint64_t size, uint64_t time)
 {
     khash_t(fih) *h = fih->h;
     int r;
